requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==4.1
django-cloudinary-storage==0.3.0
django-environ==0.10.0
django-jet-reboot==1.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
feedparser==6.0.10
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.3.0
zipp==3.15.0
